add method to get real dates from sheet
This one is a bit sad... We have to get proper Java Date instances if we want to reliably select the correct date automatically.
Until now, we only use the formatted date and that could really be anything, so it is not reliably possible to get the correct date from the string.
Instead, we need to get the correct date from a different Google Sheets API (we are using spreadsheets.values.get when we need spreadsheets.get).
That different API is a lot more complex, as it deals with GridData, but allows us to read the raw dates in SERIAL_NUMBER format (days since December 30th 1989).

All this is crazy complex to test though since all Google's classes are final and not mockable. Hence unit tests are skipped here. Feature tests will need to suffice.

[#148127607]